autoCastingdelete

+++ progressBar.cpp
@@ -13,16 +13,16 @@ progressBar::~progressBar()
 
 HRESULT progressBar::init(int x, int y, int width, int height)
 {
-	_x = x;
-	_y = y;
+	_x = (float)x;
+	_y = (float)y;
 
 	_rcProgress = RectMake(x, y, width, height);
 
-	_progressBarTop = IMAGEMANAGER->addImage("frontBar", "hpBarTop.bmp", x, y, width, height, true, RGB(255, 0, 255));
-	_progressBarBottom = IMAGEMANAGER->addImage("backBar", "hpBarBottom.bmp", x, y, width, height, true, RGB(255, 0, 255));
+	_progressBarTop = IMAGEMANAGER->addImage("frontBar", "hpBarTop.bmp", (float)x, (float)y, width, height, true, RGB(255, 0, 255));
+	_progressBarBottom = IMAGEMANAGER->addImage("backBar", "hpBarBottom.bmp", (float)x, (float)y, width, height, true, RGB(255, 0, 255));
 
 	//프로그레스 바 이미지의 크기가 가로크기가 된다.
-	_width = _progressBarBottom->GetWidth();
+	_width = (float)_progressBarBottom->GetWidth();
 
 	return S_OK;
 }
@@ -33,7 +33,7 @@ void progressBar::release()
 
 void progressBar::update()
 {
-	_rcProgress = RectMakeCenter(_x, _y, _progressBarTop->GetWidth(), _progressBarTop->GetHeight());
+	_rcProgress = RectMakeCenter((int)_x, (int)_y, _progressBarTop->GetWidth(), _progressBarTop->GetHeight());
 }
 
 void progressBar::render()
@@ -41,14 +41,14 @@ void progressBar::render()
 	//렌더는 그려주는 순서 영향을 받는다. 그려지는 순서대로 앞으로 나온다
 	IMAGEMANAGER->render("backBar", getMemDC(),
 		_rcProgress.left + _progressBarBottom->GetWidth() / 2,
-		_y + _progressBarBottom->GetHeight() / 2, 0, 0,
+		(int)_y + _progressBarBottom->GetHeight() / 2, 0, 0,
 		_progressBarBottom->GetWidth(), _progressBarBottom->GetHeight());
 
 	//앞에 그려지는 게이지의 가로크기를 조절한다.
 	IMAGEMANAGER->render("frontBar", getMemDC(),
 		_rcProgress.left + _progressBarBottom->GetWidth() / 2,
-		_y + _progressBarBottom->GetHeight() / 2, 0, 0,
-		_width, _progressBarBottom->GetHeight());
+		(int)_y + _progressBarBottom->GetHeight() / 2, 0, 0,
+		(int)_width, _progressBarBottom->GetHeight());
